# Cmp/div optimizations, tests