--- a/rebar.config
+++ b/rebar.config
@@ -1,1 +1,8 @@
 {xref_checks, [undefined_function_calls]}.
+
+{plugins, [rebar3_proper]}.
+{profiles,
+    [{test, [
+             {deps, [proper]}
+    ]}
+]}.