Чуть потюнил decimal, тесты и перфоманс замеры прилагаются.

Test source:
-----------------
```erlang
perf_test(N) ->
    Precs = [100],
    Rounds = [round_floor,round_ceiling,round_half_up,round_half_down,round_down],
    FloatList = [X/10000 + X*X || X <- lists:seq(1, N)],
    DecList = [to_decimal(F, #{precision => 100,rounding => round_half_up}) || F <- FloatList],

    Cmp3TotalTime =
    [begin
         Ctx = #{precision => P,rounding => R},
         {DivTime, _} = timer:tc(fun()-> lists:foldl(fun(D,Acc)-> cmp(Acc, D, Ctx), D end, {1,2}, DecList), ok end),
         io:format("[~p] Cmp/3 speed: ~p op/sec~n", [R, round(N / (DivTime / 1000000))]),
         DivTime
     end || P <- Precs, R <- Rounds],

    {Cmp2, _} = timer:tc(fun()-> lists:foldl(fun(D,Acc)-> cmp(Acc, D), D end, {1,2}, DecList), ok end),
    io:format("~nCmp/2 speed: ~p op/sec~n", [round(N / (Cmp2 / 1000000))]),
    io:format("Cmp speedup (%): ~p~n~n", [(((lists:sum(Cmp3TotalTime) / length(Cmp3TotalTime)) / Cmp2) - 1) * 100]),

    OldDivTotalTime=
    [begin
         Ctx = #{precision => P,rounding => R},
         {DivTime, _} = timer:tc(fun()-> lists:foldl(fun(D,Acc)-> divide_old(Acc, D, Ctx), D end, {1,2}, DecList), ok end),
         io:format("[~p] Old Div speed: ~p op/sec~n", [R, round(N / (DivTime / 1000000))]),
         DivTime
     end || P <- Precs, R <- Rounds],

    io:nl(),
    NewDivTotalTime=
    [begin
         Ctx = #{precision => P,rounding => R},
         {DivTime, _} = timer:tc(fun()-> lists:foldl(fun(D,Acc)-> divide(Acc, D, Ctx), D end, {1,2}, DecList), ok end),
         io:format("[~p] Div speed: ~p op/sec~n", [R, round(N / (DivTime / 1000000))]),
         DivTime
     end || P <- Precs, R <- Rounds],

    io:format("Div speedup (%): ~p~n", [(lists:sum(OldDivTotalTime) / lists:sum(NewDivTotalTime) -1) * 100]),
    ok.
```

Results:
------------------

[round_floor] Cmp/3 speed:              13878095 op/sec
[round_ceiling] Cmp/3 speed:            16089069 op/sec
[round_half_up] Cmp/3 speed:            15739356 op/sec
[round_half_down] Cmp/3 speed:          16010759 op/sec
[round_down] Cmp/3 speed:               12946492 op/sec

Cmp/2 speed:                            24848425 op/sec
Cmp speedup (%):                        67.69

[round_floor] Old Div speed:            1034022 op/sec
[round_ceiling] Old Div speed:          1031561 op/sec
[round_half_up] Old Div speed:          883968 op/sec
[round_half_down] Old Div speed:        879010 op/sec
[round_down] Old Div speed:             1154255 op/sec

[round_floor] Div speed:                1268828 op/sec
[round_ceiling] Div speed:              1209800 op/sec
[round_half_up] Div speed:              1044549 op/sec
[round_half_down] Div speed:            1059044 op/sec
[round_down] Div speed:                 1380506 op/sec
Div speedup (%):                        19.59